Moved main() from socket.c to main.c to make testing easier
